ai-engine/requirements.txt
pandas
scikit-learn
nump
kafka-python
